Make checkModulePrefixMismatch its own task
