+++ lib/tasks/transform-files.js
@@ -8,7 +8,6 @@ import transformTestem from '../transforms/testem.js';
 import transformTestHelper from '../transforms/test-helper.js';
 import transformApp from '../transforms/app.js';
 import { resolveWithExt } from '../utils/resolve-with-extension.js';
-import { getAppName } from '../utils/get-app-name.js';
 
 export default async function modifyFiles(options) {
   await transformWithReplace('index.html', transformIndexHTML, {
@@ -32,9 +31,6 @@ export default async function modifyFiles(options) {
     options,
   );
 
-  // Compatibility assurances
-  await handleModulePrefixMisMatch(options);
-
   try {
     let code = await readFile('.gitignore', 'utf-8');
     if (!code.match(/\/?tmp\/?[\n\r]/)) {
@@ -46,17 +42,6 @@ export default async function modifyFiles(options) {
       console.log(e);
     }
   }
-}
-
-async function handleModulePrefixMisMatch(options) {
-  const packageJSON = JSON.parse(await readFile('package.json', 'utf-8'));
-  const modulePrefix = await getAppName();
-
-  if (packageJSON.name === modulePrefix) return;
-
-  throw new Error(
-    `Unexpected modulePrefix mismatch! package.json#name is ${packageJSON.name}, but modulePrefix is ${modulePrefix}. These two values should match`,
-  );
 }
 
 async function transformWithReplace(file, transformFunction, options) {
